git todo

+++ ft_bzero.c
@@ -6,16 +6,18 @@
 /*   By: aialonso <aialonso@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/09 13:51:53 by aitor             #+#    #+#             */
-/*   Updated: 2025/10/02 18:27:26 by aialonso         ###   ########.fr       */
+/*   Updated: 2025/10/06 13:26:17 by aialonso         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 //Se encarga de rellenar un array con ceros.
 
-void	ft_bzero(void *s, unsigned long n)
+#include "libft.h"
+
+void	ft_bzero(void *s, size_t n)
 {
-	unsigned long	i;
-	char			*y;
+	size_t	i;
+	char	*y;
 
 	i = 0;
 	y = (char *)s;
