+++ ft_memcpy.c
@@ -6,17 +6,19 @@
 /*   By: aialonso <aialonso@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/09 16:00:33 by aitor             #+#    #+#             */
-/*   Updated: 2025/10/02 18:31:21 by aialonso         ###   ########.fr       */
+/*   Updated: 2025/10/06 14:34:23 by aialonso         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-//
+//Copia en destino, lo que haya en src sin importarle si se copia bien.
 
-void	*ft_memcpy(void *dst, const void *src, unsigned long n)
+#include "libft.h"
+
+void	*ft_memcpy(void *dst, const void *src, size_t n)
 {
-	unsigned long	i;
-	char			*d;
-	char			*s;
+	size_t	i;
+	char	*d;
+	char	*s;
 
 	i = 0;
 	d = (char *)dst;
