+++ ft_memmove.c
@@ -6,15 +6,19 @@
 /*   By: aialonso <aialonso@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/09 19:34:06 by aitor             #+#    #+#             */
-/*   Updated: 2025/10/02 17:08:52 by aialonso         ###   ########.fr       */
+/*   Updated: 2025/10/06 16:09:34 by aialonso         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memmove(void *dst, const void *src, unsigned long len)
+//Copia en destino, lo que haya en SRC asegurando que se copie bien.
+
+#include "libft.h"
+
+void	*ft_memmove(void *dst, const void *src, size_t len)
 {
-	unsigned long	i;
-	char			*d;
-	char			*s;
+	size_t	i;
+	char	*d;
+	char	*s;
 
 	i = 0;
 	d = (char *)dst;
