+++ libft.h
@@ -6,9 +6,11 @@
 /*   By: aialonso <aialonso@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/30 18:52:03 by aialonso          #+#    #+#             */
-/*   Updated: 2025/10/01 17:10:26 by aialonso         ###   ########.fr       */
+/*   Updated: 2025/10/06 13:25:31 by aialonso         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
+
+#include <stddef.h>
 
 int				ft_isalpha(int c);
 int				ft_isalnum(int c);
@@ -16,12 +18,12 @@ int				ft_isascii(int c);
 int				ft_isdigit(int c);
 int				ft_isprint(int c);
 int				ft_tolower(int c);
-void			ft_bzero(void *s, unsigned long n);
-void			*ft_memcpy(void *dst, const void *src, unsigned long n);
-void			*ft_memmove(void *dst, const void *src, unsigned long len);
-void			*ft_memset(void *b, int c, unsigned long len);
+void			ft_bzero(void *s, size_t n);
+void			*ft_memcpy(void *dst, const void *src, size_t n);
+void			*ft_memmove(void *dst, const void *src, size_t len);
+void			*ft_memset(void *b, int c, size_t len);
 char			*ft_strchr(const char *s, int c);
-unsigned long	ft_strlcat(char *dst, const char *src, unsigned long dstsize);
-unsigned long	ft_strlcpy(char *dst, const char *src, unsigned long dstsize);
-unsigned long	ft_strlen(const char *s);
+size_t			ft_strlcat(char *dst, const char *src, size_t dstsize);
+size_t			ft_strlcpy(char *dst, const char *src, size_t dstsize);
+size_t			ft_strlen(const char *s);
 char			*ft_strrchr(const char *s, int c);
